scanner/Scanner/Scanner/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
exurl==1.2.1
idna==3.4
python-nmap==0.7.1
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
